Update pipeline.py
added more functions to this in order to change the serve.py script so that it works faster as opposed to loading the dataset multiple times to get a quick answer. Once the application is deployed, it creates the faiss database, and then waits for a query to run the predictions; much faster.